[SQL] Reorder release patches for 24.0 release (#8390)
Extracts one patch from the main release patch file for the 24.0 LORIS release and adds it to a new file to be run ahead of any PHP scripts. this is required as all PHP scripts (even the ones that have to be run before the release patch) depend on these SQL statements causing a circular dependency.