Make getModule register autoloader for module
PR#8287 refactored the code in a way which removed the side-effect
of registering an autoloader when calling Module::factory. This was
probably not a good idea. Calls to methods on a Module instance
retrieved now crash when attempting to call a method on the module
if the module uses a class internal to the module.

Since the caller doesn't know if the module is going to use internal
namespace classes, this re-adds the implicit autoloader registration
when getting a module to the LorisInstance object.

It also caches the module descriptor in the LorisInstance after it's
retrieved in order prevent registering the same autoloader multiple times.