[SQL] 24.0.0 release patch fix truncation error (#8438)
Since the order in which the patches were run was changed by #8390. The statements in this PR which add the date_format cause a truncation error since they also "remove" the log_level enum option.

The patch that is run first now as patch _A already contains the date format and is thus sufficient to replace the statements removed here.