Validate LINST scoring scripts before running and log. (#8440)
This updates the default instrument score() function (mostly used
by LINST files) to check that the script is executable before trying
to run it. If it's not executable, it will throw a ConfigurationException
and log the reason.

Some more logs of various levels are also added to aid in diagnostics
and the print statement is replaced by an log.