[Database] remove showDatabaseQueries config.xml option (#8398)
With this the venerable debug option showDatabaseQueries is no more. It has been replaced with the Database debug log level, which now logs queries to the error log instead of printing them to the wild.

RIP showDatabaseQueries 1999(est.)-2023.

It is survived by its config.xml contemporaries sandbox, database connection information, and PSCID structure.