[mri_violations] Provide feedback on resolution status change
There is currently no feedback to the user when they change the
resolution status of a violation that something has changed, which
can be confusing.

This fades the table cell to green after the cell has been changed
in the same way that the conflict resolver does, so that the user
is aware that something happened.

Green is probably not the ideal colour, but is consistent with the
conflict resolver.