[candidate_parameters] Allow clearing of consent status (#7772)
This allows users to set a consent status back to null. I chose not to do this with a "Clear" button because a clear button would mean you are clearing all or none of the consent statuses on the page, or otherwise we would need a clear button for each individual consent. Instead, the participant can set the status to be empty and update the consent as usual. The date and withdrawal date are cleared and disabled when the status is set back to empty.

Resolves #4149