[dashboard/statistics] Reactification (#8476)
CCNA had loading issues of the dashboard because of the smarty widgets. Solved by getting the statistics data after the frontend is loaded.

Updates:

new panel component that handles multiple views.
Instead of smarty we'll use react for the statistics widgets.
Update the c3 & d3 graph generation to use the npm version.