[MRI violations] Fix hashing of MRI violation module (#8664)
In the MRI violation module, the hash generated for the violations_resolved table was created under the assumption that SeriesUID is enough to determine a unique image file. Unfortunately, this is not true anymore and multiple files with the same SeriesUID can be generated. The new unique criteria for a file is SeriesUID/EchoTime/EchoNumber/PhaseEncodingDirection combo.