[instruments] Fix LINST metadata fields not in data dictionary (#8689)
Currently, the metadata fields are only defined in the LINST instrument's data dictionary for the 'top' page, causing an error when trying to access those fields' values from other subtests. This is happening because the data dictionary definition is being called only when the fields' input elements are being added to the page in _addMetadataFields function. This changes to logic to follow the data dictionary definition for all other fields which get called whether or not the fields' input element is being added to the current page i.e. outside of if ($addElements) { } block