[LINST] Add .meta option to skip XIN rules (#8714)
XIN rules default to requiring everything unless otherwise
specified. This is not useful for all instruments (for instance,
LINST files that are used to import data from other systems where
we don't support their rule format) so this adds an option
to disable the rules for an instrument in the .meta file.