[api] Add cohorts to project info
The API documentation says you can create a visit by sending a PUT
request to the visit label including all of the visit meta data that
the endpoint returns, but there's no way to get a list of cohorts which
is part of the required data for creating a new visit.

This makes the following changes to v0.0.4-dev in order to support the
creation of visits through the API without resorting to out-of-band information:
1. Adds cohorts to /projects/$projectname (similarly to how there's a visits
   key)
2. Renames the "Battery" key to "Cohort" in v0.0.4-dev (since it's no longer
   a SubprojectID in LORIS.)

Ideally, cohorts would be more comprehensively supported by the API, but this
is the minimal that seems to be required in order to create visits through the
API and is consistent with how projects visits are handled.