[instruments] Remove CommentID from getInstanceData.
The CommentID is not part of the data, it's the foreign key used
between the flag table and the instrument table. JSON-based instruments
do not have it, and this ensures better consistency between the two
so that issues such as #8796 and #8801 will not vary based on instrument
type and will be caught sooner.

Resolves part of #8804 (Inconsistency #2)