[test] Fix order of schema sourcing
PR #8802 increased the speed of our test suite a little, and gives
a better at-a-glance overview of our API status, but the order that
`find` finds the schema patches is inconsistent and can sometimes
be sourced out-of-order, causing sporadic test failures.

This sources them explicitly instead of using find, to ensure the
order is always correct.