[battery_manager] use visit controller to fetch visit list (#8485)
The battery manager currently uses the Utility::getVisitList() function to list the visits available. That utility function queries the session table for existing visitlabels used in at least one session. That is problematic when you are creating a new visit and new instruments and want to assign instruments to the new visit, it simply does not show up in the drop down.

This PR fixes that problem and removes the use of Utility::getInstrumentList to get an accurate list of full names for instruments and get rid of deprecation warnings

Resolves #8459