[tests] Increase speed by removing duplication of work between jobs
This increases the speed of our tests by using separate jobs on GitHub Actions
for building and testing. Currently every job in our test matrix does a complete
build of all javascript, PHP, and sourcing of the database (often twice inside
of the docker image.) About 2 minutes of each test job is building LORIS.

This fixes that for the API tests so that there is 1 job to build the javascript
(once, instead of once per test job), 1 job to build the PHP (once for each version of
PHP, instead of for each version of PHP times the number of tests in the test
matrix) and has the test jobs themselves depend on those jobs and use their output
instead of duplicating the work 16 times.

Mostly, this will be noticable in combination with #8816 (which does a similar
optimization for the Raisinbread database) when building javascript (slow) and
rb (also slow) will happen in parallel.

This is only done for the API tests (which don't depend on docker compose, where
much of the work is duplicated yet again) but after #8816 should be fairly easy
to extend to the unit tests and static tests (which don't depend on Selenium.)