[Instruments] Handle empty bulkLoadInstanceData (#8857)
The logic of bulkLoadInstanceData constructs an SQL statement
with the parameters bound to it by looping over the iterable
passed to it. If the iterable is empty, the result will generate
SQL with the string `.. IN ()` which is not valid SQL. This
returns an empty array in that case before attempting to generate
the SQL, since the result of bulk loading nothing should be
nothing.