[conflict_resolver]  Add description field (#8845)
This adds a 'Description' field to the Conflict Resolver module (both Resolved and Unresolved tabs), which corresponds to the description of the 'Question' with the conflict. It uses /dictionary/module/instruments to get the description of the fields. Since some fields don't have descriptions (e.g. some _status variables), the empty string is used as default.

Changes have been made to the Instrument_LINST file such that numeric and date elements are added to the dictionary, even if they are not in the top page (inspired by the change in #8869).