[dataquery api] Clarify schema
This makes a couple updates to the schema defined for the dataquery API.

Visits may be defined for session scoped fields. A description is added,
but the field is not marked as required because it is not used for candidate
fields.

QueryResults is better defined. The separators are explicitly defined to be those
currently used by `src/Http/DataIteratorBinaryStream` and the cells are now
always defined to be JSON values. The format of session scoped data is more
explicitly defined to be in line with the expectations of the new data query
frontend.