[configuration] Escape values in config module (#8901)
PR#8759 converted the escape module to use unsafeInsert/update
to save data and prevent double escaping issues. The usages of
the textarea were audited to make sure they were properly escaped,
however the value is also displayed in the configuration module
itself. Until the module is updated from smarty to react (PR#8471),
they need to be escaped in the config module itself.

This adds escaping to the config module smarty template.