[dictionary] Remove beta tag from dictionary module
The "beta" dictionary module was added in LORIS 24, and tested in
both 24.1, and 25. When 26 is released, it will be the 4th release
including it. It's API is also used by the conflict resolver to get
field descriptions.

It should no longer be considered beta.

This removes the "beta" tag from the description, and adds a "legacy" to
the old datadict, which needs to be maintained until the new DQT is released
(based on the new dictionary.)