[script] Add script to create a new module
This adds a helper script to reduce the boilerplate of
creating a new module.

It creates the module structure and stubs for all necessary
files/directories in the module. The script takes a directory
path as an argument, verifies that it doesn't already exist and
can be written to, deduces the module name from the last part of
the path, and then creates all necessary files doing the minimum
of what needs to be done.

The module contains a root page and single endpoint named "some_api"
as an example.

If the --tsx option is specified, a stub tsx index is created as
well.

Note: SQL is contained in the module's directory rather than the
LORIS SQL directory because the script doesn't know if it's a
project or LORIS module being created (and if project, where SQL
is tracked.)