[Tools] fix timepoint data problems (#8949)
The script now uses 'Other' has Feedback type instead of the one with id 5 (Which was failing due to foreign key constraint if there was no feedback type with id = 5)

Also remove the part where the date of birth was determined based on the Subproject/Cohort id 1 or 2 (which does not mean anything) and now relies on Timepoint::getEffectiveDateOfBirth. I assumed that subproject 1 and 2 was making a distinction between pre ans post natal visits. Timepoint::getEffectiveDateOfBirth should handle that.

Resolves #5040