[DataTable] Fix hide downloadCSV option (#8995)
When hide downloadCSV is set to true the Download CSV option
is still shown in the table header, but hidden from the table
footer. This fixes it so that it is hidden from both.

Hide downloadCSV also currently incorrectly hides the footer. Fix so that
only the CSV button is hidden.