[dataquery] New Data Query Module (#8907)
This implements a completely new, CouchDB free data query module based on the LORIS data query framework and the API introduced in PR#8268.

In the new module, all queries that are run by the user are saved into a history. The user can star or name a query in order to make it easier to find, rather than having to decide ahead of time that they want to save or share a query. Admins can pin queries (either to the top of the module or to the LORIS dashboard).

The UI is also (generally) simplified and more context-sensitive. The "define filters" tab should be easier to understand. The view data tab has a new view where candidates are rows, fields are columns, and session-scoped variables are displayed inline in the cell in a table.

Permissions are enforced for modules, candidates, and sessions.