[api/instruments] Add endpoint that allows you to add instruments to a battery (#8868)
This implements a /candidates/{id}/{visit}/instruments POST endpoint which allows you to programmatically add an instrument to a candidate's battery through the LORIS API.

The POST endpoint takes input in the same format as GET, but adds any missing instruments to the battery instead of simple returning the existing test battery. It returns the new battery.