[api] Include filename in error message (#8776)
There is an error message in the web server logs when a file is in the database but not on the filesystem but it doesn't tell you where it was looking on the filesystem. This adds the full path to make it easier to determine where LORIS is trying to find the file when ie. setting up raisinbread.