[jsx] Remove DynamicDataTable (#8993)
DynamicDataTable is an old wrapper around StaticDataTable which fetched
the data and then loaded the StaticDataTable. This has entirely been
replaced by other data table types. The only remaining references are in
deprecated wiki and candidate_list/onLoad.js. onLoad.js is also not used
anymore (it is compiled, but it isn't loaded in the getJSDependencies of
the page.)

This removes the dead code of DynamicDataTable and onLoad.