[my_preference] Add support for my_preference widgets (#8938)
This adds a new widget type, "userpreference" which allows modules to plug content into the user's my_preferences page.

It works similarly to the candidate_profile widget, where it assumes each call will return an array of UserPreferenceWidget's to add to the page, each one denoting a single react component. The UserPreferenceWidget must have:

- A title to display as a heading
- A URL to get the javascript for the component
- The component name to instantiate
- React props to pass to the widget

This is useful for ie. the OIDC module adding a widget where the user can link an external openid account to the logged in user, but can also be used for any other modules which have preferences that should be configurable by the user on the my_preferences page.