Use dobFormat config more places (#8937)
This updates a few more places to use the dobFormat LORIS config.

It adds the getDisplayDoB function proposed in #8935 (Resolves #8935) and uses it on the timepoint list and instrument list header tables. It also adds support for the config variable (without using the helper function) in the candidate_list page. (There is no access to the candidate object in the provisioner and it needs to be refactored to use the CandidateData DTO with the Candidate constructor efficiently in order to have access to it, but in the meantime this honours the variable.)