[issue_tracker] Fix inactive user query (#8904)
This fixes the logic in the inactive users query for the assignee drop down of the issue tracker.

The query was missing brackets in the logic and was querying active users in the inactive users array. This resulted in all users being unset from the assignees object for users who do not have the "access_all_profiles" permission. With brackets, users without the "access_all_profiles" permission should be able to see the appropriate users in the assignee drop down of a new issue.