[htaccess]  Modifies rule to allow spaces in filename when downloading files. (#9029)
The flags B and BNP were added to the rewrite rule that redirect to index.php. This is need as files who have spaces in the name are not longer downloading (e.g. the Document Repository)

This issue seems to appear after the upgrade of the Apache server. More details can be found in the original issue.

Resolves #8706