StaticDataTable mixed column types sorting (#9120)
When mixing numbers with strings in an instrument field value (e.g. text field) results in a DQT column that has a sorting issues. This happened for a "summary" instrument in HBCD project that import results from multiple other instruments. This instrument has to clearly display the value of the field (i.e. numeric) or "No data" (i.e. string).