Make candidate_list use LORIS streamable binary format
This updates the candidate_list module to use format=binary
format instead of format=json. For large projects with many
candidates the encoding of the data to JSON can use up all
the server's memory and then crash with an out-of-memory error.

Since the format=binary format is a line-based streamable format,
it does not require loading the entire result set into memory before
transmitting it to the client and larger data sets can be loaded into
LORIS.