[statistics] Optimize recruitment graph queries
The recruitment graphs on the dashboard provided by the
statistics module are not efficient. They run one SQL query
per month per year per site for each graph that is displayed.
This is highly inefficient for large projects.

This replaces the single queries in the inner loops with a
single group-by query which it processes into the same format.

The result should be significantly faster load times on the dashboard
when the statistics module is loaded.