[publication] Fix number of files in Upload tab (#9179)
In the uploadForm.js file is modified to increase / decrease the numFiles value according to whether a file was actually selected or not. If a file is selected, the numFiles increases. If a file is not selected, the numFiles decreases.

This prevents the bug where "File to Upload" field is added indefinitely if the file is browsed and then cancelled.

Fixes #9136