[SQL] Move flag.data column to own table
This moves the data from flag.Data to a new table. The large size
of the data was causing scans of the flag table to take a significant
amount of time on large datasets due to fragmentation, even if the
column was not used. The large amount of was was slowing
down any query that joined the table (even if indexes were used.)

Moving the data to a new table means that queries that do a table
scan are faster since joining the new instrument_data table only needs
to pull in data that it needs.

(On a database with 6 million rows in flag, this change brought the
timing of the query used for bulkLoadInstanceData down from ~5s to ~1.5s
when bulk loading all the data for a single instrument.)