Cache module dictionary staticly in Query class (#9217)
The Query class in the dataquery module includes an IsAccessibleBy function to check accessibility of a query by checking if the user has access to all fields used in the query (for both fields and filters). It currently caches the data dictionary of a module the first time it's used to avoid having to re-retrieve the dictionary for every field. However, each Query object instance has its own cache.

This changes the dictionary cache to be a static variable so that it can be shared across different instantiations of the object. This significantly speeds up the /dataquery/queries/ endpoint which applies an AccessibleResource filter to every query returned by a provisioner.