Fix unused vars eslint warnings (#9226)
Fix errors unused-vars warnings that are easily fixed errors. I'm not
sure of the impact of changing the remaining two since they're used in
a destructuring context. Someone familiar with the EEG browser should fix
(and test) them.

Upgrade rule from "warning" to "error" for javascript, while leaving it for
typescript since there are two that are still unresolved.