[dataquery] Fix strikethrough effect when field re-added
When a field is removed by the trashbin icon, the onMouseLeave
for the hover effect doesn't fire as the component is unmounted.
When a new field is added, the wrong "hover" index is set and it
incorrectly shows a strike-through effect.

This is effectively the same fix as #9104 on a different tab.

Resolves #9249