[dataquery] Filter candidates from SQL instead of PHP
This updates the bulkLoadInstanceData function to optionally
take a single QueryTerm to load candidates for, in addition
(or instead of) a list of CommentIDs. If provided, the loading
will filter for instruments matching the term from SQL, rather
than loading them all into memory and filtering from the PHP
side.

This should significantly reduce the memory consumption for
large data sets.