[dataquery/instruments] Handle cardinality::many instruments in DQT
The ability for the frontend to display session variables with a cardinality
of "many" per session was added for the imaging, but the instrument queryengine
was not updated to properly format multi-select elements. This adds support
for multiselect elements by updating the instrument queryengine to return
correctly formatted data for instruments.