[Core] Add database query class to represent the results of a DB query (#9334)
This adds `\LORIS\Database\Query` class which is both an iterator an
countable to replace the return value of `Database->pselect()`. This
improves the memory consumption of large datasets, because the iterator
only needs to load the row that is being foreach'd over, instead of
loading the entire query result into memory.

Places that currently depend on it being an array are wrapped in
`iterator_to_array` (They are mostly queries returning small results, so
this isn't expensive. The optimization is mostly useful for places like
the dataquery module which iterate over large amounts of data.)