[Core] Add CenterID singleton (#9337)
This adds a singleton method to the CenterID to cache the instantiation and re-use it. The value of a CenterID often comes from the database as an int, but needs to be converted to a CenterID class in order for various function calls and type checking (in particular, for permission access.) Instantiating a new CenterID each time uses a large amount of memory and puts a large degree of pressure on the garbage checker for big datasets, especially places that need to check access to many candidates or timepoints (such as the dataquery module.)

Replacing it with a singleton uses much less memory in these cases.