[dataquery] Optimize filtering of valid candidates for large datasets
This optimizes the filtering of inaccessible candidates, building on
PR#9334

Previously, the entire query was loaded into memory, and then manipulated
in a way that made it easier to see which candidates are valid, then returned
as an array by putting the valid ones in a temporary table and selecting it.

With this change, instead of an array, a generator is used so that only one
candidate (and its sessions) need to be loaded into memory at a time. This is
done on an unbuffered connection, which necessitates a different database
connection since only one query can be run at a time on a connection and the
caller may make other queries before the generator finished.