[Optmization] Replace bulkLoadInstanceData with generator
BulkLoadInstanceData currently returns an array of instruments
for each candidate. This is inefficient because every instrument
needs to be loaded into memory, when the results are only ever
used in a foreach loop. This replaces the array_map with a generator
that yields the instrument, and uses an unbuffered database connection.

This uses significantly less memory for large datasets as now only one
CommentID needs to be loaded into memory at a time.