[Memory] Remove unnecessary iterator_to_array
The InstrumentQueryEngine converts the results of a query
that returns all values from an iterator to an array because
it's used twice. However, the second usage is unnecessary as the
data is available in another way. This switches the second usage
to be array_keys of a variable created in the first iteration and
removes the iterator_to_array, which should result in less memory
usage as it can be freed after the first iteration.