[Utility] Remove unnecessary joins in queries
There were (at least) 2 unnecessaries joins in database queries
in getVisitList and getVisitInstruments where a table was joined,
and then the table was only used to filter for the same key that
was used in the join condition.

Remove those joins by just filtering on the key in the first place.