[dataquery/instruments] Handle cardinality::many instruments in DQT (#9305)
The ability for the frontend to display session variables with a
cardinality of "many" per session was added for the imaging, but the
instrument queryengine backend was not updated to properly format
multi-select elements. This adds support for multiselect elements by
updating the instrument queryengine to return correctly formatted data
for instruments with the many cardinality.