[candidate_list] Fix 'Access Profile' link on timepoint list page (#9455)
Selection Filter boxes on 'Access Profile' page are
empty when accessed via breadcrumb link.
When the 'Access Profile' is reached via clicking on the breadcrumb at, the
page will load, but the multi-select boxes in the 'Selection Filter'
Menu appear empty. This is not the case if the page is accessed via the
link in the 'Candidate' menu.