[Core] Fix loading of pselectRow queries
Queries which use pselectRow that return data can fail to load with
an error about not being able to execute queries with pending results.

pselectRow does a count on the database connection, which executes
the query. When getFirstRow is called after, the query has already
been executed but the results have not been fetched when getIterator()
attempts to re-execute the query.

This keeps track of the execution state so that it is only executed
the first time.