[Core] Faster php syntax checking
PHP 8.3 added the ability to pass more than one filename at a
time to 'php -l'. As such, we no longer need to specify that xargs
should only lint one file at a time and let it work on batches now
that the minimum supported version of LORIS is PHP 8.3.

This improves the time of running 'make checkstatic'.