[Candidate] Support creating candidates from the backend (unix user) (#9485)
The current code tries to get a username from the session. the session
does not have a state when the candidate::createnew function is called
from a script on the backend.

The new code is inpired by the database class which tries to use the
environment username to create the data.

This allows candidate::createNew to be called from scripts.