[Cleanup] Make DBRowProvisioner use Database class (#9509)
This updates DBRowProvisioner to remove the strval (and fixes #9335).

After doing that, it became clear that the inner class could be replaced
by a generator while still fulfilling the contract of returning a
`\Traversable`. Doing this made it more obvious that since #9334,
pselect could be used directly while maintaining the lazy evaluation
instead of bypassing the Database object to use the PDO directly.
However this required injecting a `LorisInstance` object to the
constructor to get the database connection.

With access to the `LorisInstance` object, the `DBRowProvisioner` can
also get a new database connection and disable query buffering to lazily
retrieve rows since PR #9344. This may theoretically improve performance
and memory usage on large provisioners but has not been tested.
(Performance did not decrease in raisinbread but it is not a large
dataset where it would be noticeable.)