Move AWS to middleware
This refactors the way AWS S3 buckets are handled so that the
stream wrapper registration is added by a middleware. For scripts,
the logic of registering a stream wrapper is moved to generic_includes.

This is a first pass at addressing the comment in htdocs/index.php that
most of the work of NDB_Client should be done by middleware.