[SQL] Refactor FK CandID int(6) to CandidateID int(10) (candidate.ID PK) (#9556)
This refactors the DB schema to rename every table except
`candidate`'s `CandID` field to `CandidateID`. This field now instead
references `candidate`.`ID`, which is now the PK of the table and a
`int(10)` instead of `int(6)` previously. `candidate`'s `CandID` field
has also been upgraded to an `unsigned int(10)`.

There is a patch that sets all the values of `CandidateID` of all tables
to instead become the corresponding `candidate`.`ID`.

The back-end of most modules that used a table with a `CandID` were
refactored to use `candidate`.`CandID` and to instead `JOIN ON
candidate.ID = table_with_CandID_previously.CandidateID`.

The result is that CandIDs have 1 canonical source now, and changes
(such as increasing its length) do not need to be made to every foreign
key.

Fixes #8827