[Login] Summary Statistics (#9518)
- Added a customizable summary of LORIS statistics to the Login page for
guest users to see
- A new table was created that is populated with the counts of pinned
dataquery queries, and separated by project
- This table is read by the login module for projects with
`showSummaryOnLogin` as true in the Project database table.
- If no data is populated, then the summary is not shown on the login
page
- Data is populated by the new tool:
`tools/update_login_summary_statistics.php`. It must be run for the data
in the summary to be updated. Can be added to a project's nightly cron,
or at each release
- Some queries are not necessarily possible in the DQT, so in-order to
keep it customizable for projects, I added the ability to add SQL files
to a folder in project/tools/Login_Summary_Statistics that can be added
to the summary list.
- There is a clipboard icon at the bottom right where if you press it,
your clipboard will be filled with a csv version of the data