[Login Statistics] Fix query
There was a semantic conflict between PR# and one of the queries
in PR#9518, as the CandID now needs to come from the candidate table.

This adds a JOIN from the session for the one query missing it, so
that the update login statistics script doesn't fail with an SQL
error.