[NDB_BVL_Instrument_LINST] Add Metadata Fields after adding to dictionary (#9667)
- raisinbread BMI instrument was not opening due to a change I made in
#9416 where getExaminerFields in NDB_BVL_Instrument now relied on data
being present in the instrument's dictionary.
- Unfortunately, when loading a LINST instrument with Direct Entry
enabled, it would first run _addMetadataFields, and then add metadata
fields to the dictionary
- _addMetadataFields would then call getExaminerFields causing it to
fail, since getExaminerFields relies on the dictionary having metadata,
but the metadata is added after _addMetadataFields
- The change has been made so _addMetadataFields is run _after_ metadata
is entered in the instance's dictionary. Further, metadata fields are
only added to the instances dictionary if direct entry is not enabled.
Otherwise they were added to the dictionary when they did not exist.

* Resolves #9582, #9644