[imaging_uploader] Site&Project check with config setting and permission (#9533)
These changes offer a backwards compatible way of adding Site and Project permissions to the imaging_uploader module.

Since the imaging_uploader may contain uploads which fail validation (for whatever reason) and are thus not associated to a sessionID in the mri_upload table, the files are not technically associated with a site and project. In the current implementation all users with access to the module have access to all uploads which is problematic for institutional instances like CBIG. The changes here include a config setting to enable/disable the use of site/project checking (disabled is the status quo) but also include a permission which can give a user access to only their site's/project's data + any file with no sessionID and thus no site or project.